requirements.txt
requests>=2.23.0
python-dateutil>=2.8.1
rpi-backlight>=2.1.0